server/requirements.txt
fastapi
qdrant_client
uvicorn
requests
python-dotenv
openai==1.55.3
pydantic==2.7.1
